massive-analytics-dashboard-duckdb-parquet/pyproject.toml
[project]
name = "massive-analytics-dashboard-duckdb-parquet"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.5.0",
    "mysql-connector-python>=9.6.0",
    "psycopg[binary]>=3.2.0",
    "tabulate>=0.10.0",
]
